Close open upvalues
This is a pre-requisite for closures. We compile an access to
a variable of an enclosing function as a GETUPVAL instruction and
mark the variable as an open upvalue of the enclosing function.

Blocks that create an upvalue that will be open at some point have
to emit code to close the upvalue (both at the end of the block and
and at any break/continue) by setting the A part of a JMP instruction
to the number of upvalues that need to be closed. The one exception
is the function block itself, since the function return always closes
all open upvalues of the function.